add argument and maintenence runs

+++ mege_all_best_folders.py
@@ -10,15 +10,15 @@ if __name__ == '__main__':
         creat_folder_flag=True
 
         for f in os.listdir(os.path.join(MODELS_DIR,i)):
-            if 'best' in f:
-                print(f)
-                for k in os.listdir(os.path.join(MODELS_DIR,i,f)):
-                    cur_path = os.path.join(MODELS_DIR,i,f,k)
-                    if os.path.isdir(cur_path):
-                        for h in os.listdir(cur_path):
-                            shutil.move(os.path.join(cur_path,h),os.path.join(MODELS_DIR, i, f))
-                        if len(os.listdir(cur_path))==0:
-                            os.rmdir(cur_path)
+            if 'best' in f and 'temp' in f and os.isdir(os.path.join(MODELS_DIR,i,f)):
+                # for k in os.listdir(os.path.join(MODELS_DIR,i,f)):
+                cur_path = os.path.join(MODELS_DIR,i,f)
+                dest=cur_path.replace('temp','')
+                #     if os.path.isdir(cur_path):
+                #         for h in os.listdir(cur_path):
+                shutil.move(cur_path,dest)
+                if len(os.listdir(cur_path))==0:
+                    os.rmdir(cur_path)
         # for f in os.listdir(os.path.join(MODELS_DIR,i)):
         #     if 'best' in f and 'temp' in f:
         #         print(f)
